Add files via upload
This is the final course assignment of the course "Introduction to Data Analytics for Business" - Coursera. Here, I have created Conceptual Business Model and Relational Data Model (subset of the conceptual model) of Insurance Industry. Here, I have identified the primary and foreign key for each data table and the type of source systems from where the data is collected. Answered related business queries with help of SQL queries. Also, identified the data sensitivity and data quality issues of the designed data model.